requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.4
blis==0.7.5
boto3==1.20.26
botocore==1.23.26
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.11.0
cymem==2.0.6
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
ec2-hibinit-agent==1.0.0
en-core-web-lg==3.2.0
entrypoints==0.3
ffmpeg-python==0.2.0
fonttools==4.28.5
future==0.18.2
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
hibagent==1.0.1
httplib2==0.20.2
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==0.10.0
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.2
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.5.1
more-itertools==4.2.0
murmurhash==1.0.6
netifaces==0.10.4
nltk==3.6.7
numpy==1.22.0
oauthlib==3.1.0
packaging==21.3
pandas==1.3.5
pathy==0.6.1
pexpect==4.6.0
Pillow==8.4.0
preshed==3.0.6
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pydub==0.25.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
PyYAML==5.3.1
regex==2021.11.10
requests==2.26.0
requests-unixsocket==0.2.0
rsa==4.8
s3transfer==0.5.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
sklearn-contrib-py-earth==0.1.0+1.gdde5f89
sklearn-pandas==2.2.0
smart-open==5.2.1
sos==4.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SpeechRecognition==3.8.1
srsly==2.4.2
ssh-import-id==5.10
systemd-python==234
thinc==8.0.13
threadpoolctl==3.0.0
tqdm==4.62.3
Twisted==18.9.0
typer==0.4.0
typing-extensions==4.0.1
ubuntu-advantage-tools==27.4
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.7
wadllib==1.3.3
wasabi==0.9.0
zipp==1.0.0
zope.interface==4.7.1
